+++ src/actions/index.js
@@ -1,7 +1,0 @@
-import startApp from './startApp';
-import stopApp from './stopApp';
-
-export default {
-  startApp,
-  stopApp,
-};